[Feature] added main shell functions - added TODO lists - added verification functions in SSD
